projects/opendiscourse_automation_bundle/requirements.txt

requests>=2.32.0
PyYAML>=6.0.2
python-slugify>=8.0.4
